requirements.txt
# pip install -r requirements.txt

requests==2.28.2
python-dotenv==1.0.0
rich==13.3.3
sentence-transformers==2.2.2
google-generativeai==0.1.1
youtube-transcript-api==0.5.0
numpy==1.25.0